refactor: Extract common functions into the common package
